refactor(database)!: revamp component & schema internals
Refactors internal component and schema representations to simplify metadata handling, improve type safety, and reduce indirections.

- Introduces a clear component_meta structure and moves vtable logic into a dedicated vtable implementation.
- Replaces the previous component index/version representations with narrower integer types and a dedicated component_version alias.
- Removes the old component_bitset alias and uses the dynamic bitset type directly where needed.
- Reworks schema and schema_builder to store component hashes, metas and vtables sorted by hash rather than previous id vectors.
- Simplifies environment storage: renames slot structure, uses dense vectors with sparse index mapping, and adjusts slot indexing logic.
- Updates table signature to construct the dynamic bitset directly.
- Updates tests to use new factory functions and types and to reflect API changes.

Why: clarify and consolidate component metadata and vtable responsibilities, reduce indirection and allocation noise, and pave the way for more efficient schema/component operations.

BREAKING CHANGE: component index/version types and component metadata APIs have changed (smaller integer widths and new meta/vtable interfaces). Callers that relied on previous component_index/component_meta/component_version layouts or helper aliases must be updated accordingly.

+++ include/ant/database/detail/component_index.hpp
@@ -6,6 +6,6 @@
 
 namespace ant::detail {
 
-using component_index = basic_index<struct component_index_tag, std::uint32_t>;
+using component_index = basic_index<struct component_index_tag, std::uint16_t>;
 
 } // namespace ant::detail
